+++ macros/src/items/timestamp.rs
@@ -29,13 +29,13 @@ pub(crate) fn expand(args: TokenStream) -> TokenStream {
     );
 
     let var_name = format_ident!("S");
-    let var_item = construct::static_variable(&var_name, &format_string, "timestamp");
+    let var_item = construct::static_variable(&var_name, &format_string, "timestamp", None);
 
     quote!(
         const _: () = {
             #[export_name = "_defmt_timestamp"]
             #[inline(never)]
-            fn defmt_timestamp(fmt: ::defmt::Formatter<'_>) {
+            fn defmt_timestamp(fmt: defmt::Formatter<'_>) {
                 match (#(&(#formatting_exprs)),*) {
                     (#(#patterns),*) => {
                     // NOTE: No format string index, and no finalize call.