+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - [#847]: `decoder`: Fix log format width specifier not working as expected
+- [#845]: `decoder`: fix println!() records being printed with formatting
 
 [#847]: https://github.com/knurling-rs/defmt/pull/847
+[#845]: https://github.com/knurling-rs/defmt/pull/845
 
 ## [v0.3.8] - 2024-05-17