+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [#xxx]: `CI`: Update mdbook to v0.4.40
 - [#847]: `decoder`: Fix log format width specifier not working as expected
 - [#845]: `decoder`: fix println!() records being printed with formatting
+- [#843]: `defmt`: Sort IDs of log msgs by severity to allow runtime filtering by severity
 
 [#xxx]: https://github.com/knurling-rs/defmt/pull/xxx
 [#847]: https://github.com/knurling-rs/defmt/pull/847
 [#845]: https://github.com/knurling-rs/defmt/pull/845
+[#843]: https://github.com/knurling-rs/defmt/pull/843
 
 ## [v0.3.8] - 2024-05-17